測試完成，新增到 putUserInfo 路由
